entity 설정
