--- src/main/java/com/payhub/mamber/entity/Member.java
@@ -1,29 +1,0 @@
-package com.payhub.mamber.entity;
-
-import com.payhub.order.entity.Order;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter @Setter
-public class Member {
-
-    @Id @GeneratedValue
-    @Column(name = "member_id")
-    private Long memberId;
-
-    private String name;
-
-    @Embedded
-    private Address address;
-
-    @OneToMany(mappedBy = "member")
-    private List<Order> orders = new ArrayList<>() ;
-
-    private String phone;
-
-}
